requirements.txt
# Google Gemini SDK
google-generativeai>=0.8.0

# Document parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# Database
sqlite3

# Embeddings and vector search
faiss-cpu>=1.8.0
numpy>=1.26.0

# API layer
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0
